feat: continue to refactor project structure
