+++ include/core/common.h
@@ -18,4 +18,6 @@
 
 #define UINT8_COUNT (UINT8_MAX + 1)
 
+#define MAX_IF_BRANCHES 56
+
 #endif
