+++ src/parser/expressions.c
@@ -1,0 +1,6 @@
+/**
+ * @file expressions.c
+ * This file contains the methods for parsing expressions in the CSLO language.
+ * It includes methods for parsing various types of expressions such as literals,
+ * variables, and control flow statements.
+ */
